Sorted components in appropriate folders
